# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Disk": {
				"properties": {
					"index": {
						"type": "number",
						"format": "double"
					},
					"model": {
						"type": "string"
					},
					"serialNumber": {
						"type": "string"
					},
					"firmwareRevision": {
						"type": "string"
					},
					"sizeBytes": {
						"type": "number",
						"format": "double"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"lastErrorCode": {
						"type": "string",
						"nullable": true
					},
					"errorCleared": {
						"type": "string",
						"nullable": true
					},
					"errorDescription": {
						"type": "string",
						"nullable": true
					},
					"errorMethodology": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"index",
					"model",
					"serialNumber",
					"firmwareRevision",
					"sizeBytes",
					"temperature",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partition": {
				"properties": {
					"diskIndex": {
						"type": "number",
						"format": "double"
					},
					"index": {
						"type": "number",
						"format": "double"
					},
					"bootable": {
						"type": "boolean"
					},
					"primaryPartition": {
						"type": "boolean"
					},
					"volumeName": {
						"type": "string"
					},
					"sizeBytes": {
						"type": "number",
						"format": "double"
					},
					"freeSpaceBytes": {
						"type": "number",
						"format": "double"
					},
					"usagePercentage": {
						"type": "number",
						"format": "double"
					},
					"volumeLetter": {
						"type": "string",
						"nullable": true
					},
					"fileSystem": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"diskIndex",
					"index",
					"bootable",
					"primaryPartition",
					"volumeName",
					"sizeBytes",
					"freeSpaceBytes",
					"usagePercentage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStorageInformation": {
				"properties": {
					"machineId": {
						"type": "string"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"disks": {
						"items": {
							"$ref": "#/components/schemas/Disk"
						},
						"type": "array"
					},
					"partitions": {
						"items": {
							"$ref": "#/components/schemas/Partition"
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"machineId",
					"timestamp",
					"disks",
					"partitions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_StorageInformation.Exclude_keyofStorageInformation.machineId__": {
				"properties": {
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"disks": {
						"items": {
							"$ref": "#/components/schemas/Disk"
						},
						"type": "array"
					},
					"partitions": {
						"items": {
							"$ref": "#/components/schemas/Partition"
						},
						"type": "array"
					}
				},
				"required": [
					"timestamp",
					"disks",
					"partitions"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_StorageInformation.machineId_": {
				"$ref": "#/components/schemas/Pick_StorageInformation.Exclude_keyofStorageInformation.machineId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "X-Token",
				"in": "header",
				"description": "JWT token sent via X-Token header. Example: \"X-Token: {token}\""
			}
		}
	},
	"info": {
		"title": "api-tor",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/monitoring/storage-information/machine/{machineId}": {
			"get": {
				"operationId": "GetStorageInformationByMachineId",
				"responses": {
					"200": {
						"description": "List",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IStorageInformation"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Storage Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateStorageInformation",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Storage Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_StorageInformation.machineId_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteStorageInformationByMachineId",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"deletedCount": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"deletedCount"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Storage Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/storage-information/{id}": {
			"get": {
				"operationId": "GetStorageInformation",
				"responses": {
					"200": {
						"description": "Storage Information",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IStorageInformation"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Storage Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteStorageInformation",
				"responses": {
					"200": {
						"description": "Deleted"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Storage Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/storage-information/machine/{machineId}/latest": {
			"get": {
				"operationId": "GetLatestStorageInformation",
				"responses": {
					"200": {
						"description": "Latest Storage Information",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IStorageInformation"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Storage Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/storage-information/machine/{machineId}/batch": {
			"post": {
				"operationId": "CreateStorageInformationBatch",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"count": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"count"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Storage Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/Omit_StorageInformation.machineId_"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/monitoring/storage-information/machine/{machineId}/range": {
			"get": {
				"operationId": "GetStorageInformationByDateRange",
				"responses": {
					"200": {
						"description": "Storage Information",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IStorageInformation"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Storage Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}